Prepare Filter & Aggregation move to plugin
This is the initial work to render different condition forms from
plugins.
