+++ graylog2-web-interface/src/components/event-definitions/event-definition-form/event-definition-types/index.js
@@ -1,0 +1,13 @@
+import { PluginManifest, PluginStore } from 'graylog-web-plugin/plugin';
+
+import FilterAggregationForm from './FilterAggregationForm';
+
+PluginStore.register(new PluginManifest({}, {
+  eventDefinitionTypes: [
+    {
+      type: 'aggregation-v1',
+      displayName: 'Filter & Aggregation',
+      formComponent: FilterAggregationForm,
+    },
+  ],
+}));
